+++ inst/shinyapp/engine/tabulate_extra_table.R
@@ -90,7 +90,7 @@ out <-
     results1 |> 
     dplyr::mutate(
       value_f_org = dplyr::case_when(
-        is.nan(value) ~ "-",
+        mics_missing_value_dash(value, stat_type) ~ "-",
         stat_type %in% c("n", "n_unw", "n2", "n_unw2", "hhmembers") ~ as.character(format(round(value, 0), big.mark = ",")),
         stat_type %in% c("p", "p_unw", "p(100)", "p_unw(100)", "mean", "mean_unw") ~ as.character(round(value, 1)),
         TRUE ~ as.character(value)
@@ -104,6 +104,7 @@ out <-
     out <- out |>
     dplyr::mutate(
       value_f = dplyr::case_when(
+        mics_missing_value_dash(value, stat_type) ~ "-",
         (is.na(n_unw) | n_unw == 0) & stat_type == "100" ~ "0",
         stat_type %in% c("n", "n_unw", "100") ~ as.character(value),
         (is.na(n_unw) | n_unw == 0) & stat_type != "100" ~ "-",
@@ -115,6 +116,7 @@ out <-
     ) |> 
        dplyr::mutate(
       value_f_view = dplyr::case_when(
+        mics_missing_value_dash(value, stat_type) ~ "-",
         stat_type %in% c("n", "n_unw") ~ value_f_org,
         (is.na(n_unw) | n_unw == 0) & stat_type == "100" ~ "0",
         (is.na(n_unw) | n_unw == 0) & stat_type != "100" ~ "-",